requirements.txt
asgiref==3.11.1
certifi==2026.4.22
charset-normalizer==3.4.7
Django==6.0.4
djangorestframework==3.17.1
Faker==40.15.0
idna==3.12
psycopg2-binary==2.9.12
python-decouple==3.8
requests==2.33.1
sqlparse==0.5.5
tzdata==2026.1
urllib3==2.6.3
whitenoise==6.12.0
